refactor(pixelflow-ir): move the lowering passes out of backend/emit
The lowering passes were never part of a backend

`lowering.rs` sat in `backend/emit/` next to the instruction encoders, and
nothing about it belongs there: its entire import list is `crate::{arena,
kind, variance}` plus `alloc`, and its only `cfg` is `#[cfg(test)]`. It does
not know which ISA it is feeding, and the legal set it lowers toward is the
same on every target.

The tree had already noticed. Five crates reached *past* the codegen
namespace to get at it — `pixelflow_ir::backend::emit::lowering::lower_dwrt_owned`
appears in pixelflow-search's runtime, pixelflow-compiler's ir_bridge and its
parity tests, and two pixelflow-graphics golden tests. A path that every
external caller has to apologize for is a filing error.

So: `src/passes.rs`, a sibling of `arena` and `kind` rather than a child of
the emitters. The move is mechanical — `git mv` plus the import paths — but it
is a prerequisite for anything that wants the transforms without the machine
code, including the `no_std` question, since the passes have no OS or libc
dependency and the emitters carry all of it.

The module doc described only transcendental expansion, which is one of the
four passes it holds. It now names all four with what each consumes and
produces, and records why the order is an order: differentiating a `sin`
manufactures a `cos`, so `lower_dwrt` must precede the pass that expands
them. That constraint was previously spelled out in a comment on one call
site, three of which existed.
